Fixing AppointmentService and backporting methods.
Backporting the methods we need to make the appointment service work for
module writers (such as telehealth).

We had to bring in the UuidRegistry changes to make both encounters,
users, and appts work properly.

Several methods were backported in the appointment service that relied
on newer FHIR functionality. Refactored the methods to use older
functionality so this will still work in 6_0_0.

Brought back the hasData() method in ProcessingResult (and other new
functions) to make this work.