Callback participant, force encounter date flag
Added an office contact setting in care coordination.  Added the
participants piece to the header.

Removed the fDate formatting on the document timestamp as PHP was
already sending a timestamp formatted document and the fDate was being
corrupted.  Had to do this to pass ETT validation.

Added a checkbox to let people force the author timestamp value to be
the encounter date time stamp.  This made it possible to override the
timestamps for the document across the three different test scenarios
without having to change it for each individual patient.  Most people
will have this flag turned off, but it sped up the ONC testing instead
of having to set the timestamp on each doc generation (for both b1 and
g9 testing).