Exceptions instead of die, additional session vars
Added additional session variables that were missing.  Also changed the
die logic to use exceptions so we can both log it in the database and
redirect back to the login page for a better UX experience.

Until the core PR with the config gets merged in, defaulted the
telehealth config to be enabled for one time logins.

Changed the method signature to return the auth array so we can log the
pid in the one time usage.  The service redirects the browser and relies
on the caller to exit once they've done any processing logic.