Added generate invite link feature, fixed settings.
Implemented a generate invite link feature since we have to dynamically
generate an invite link now for the one time login links.  Since the
links are generated on the backend server we have to have a two step
process for grabbing the session link.  First the link is generated and
inserted into the DOM and then the browser can copy the link and the
invitation.  Firefox and safari block the copy feature when the copied
content is generated asynchronously.

Fixed the settings stuff that got lost during the merge conflict.

Added a setting option to configure the expiration time on the the one
time links.  Has a minimum of 1 minute and a maximum of 30 minutes.  If
the setting is missing or invalid it defaults to 15 minutes.

Fixed the display of the request media permissions prompt.

Fixed some javascript errors when the conference room shuts down.

Fixed bug with the plain text email invitations missing the join link.